Refactor packets
Move the classes that extend Packet into their own files.
