+++ src/packet/HeaderPacket.java
@@ -1,0 +1,15 @@
+package packet;
+
+class HeaderPacket extends Packet {
+
+    private String fileName;
+
+    protected HeaderPacket (Byte fileId) {
+        super(fileId);
+    }
+
+    protected HeaderPacket (Byte fileId, String fileName) {
+        super(fileId);
+        this.fileName = fileName;
+    }
+}
